Fixing compile failure for 32-bit systems (#90)
* Fixing integer overflow issue for 32-bit systems

Previously `kv.go` and `value.go` were using int instead of int64
for log file size, which caused compile time issues like:

```
kv.go:139: constant 2147483648 overflows int
```

This changeset fixes the issue by using int64.

* Reverting README.md

* Reverting unexpected comment change

+++ kv.go
@@ -65,7 +65,7 @@ type Options struct {
 	ValueGCRunInterval time.Duration
 
 	// Size of single value log file.
-	ValueLogFileSize int
+	ValueLogFileSize int64
 
 	// The following affect value compression in value log.
 	ValueCompressionMinSize  int     // Minimal size in bytes of KV pair to be compressed.
